requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
Click==7.0
configparser==3.5.0
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
gunicorn==19.9.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.15.4
Pillow==5.3.0
pylint==1.9.3
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
